requirements.txt
asn1crypto==0.24.0
certifi==2019.3.9
cffi==1.12.3
cryptography==2.6.1
future==0.17.1
h5py==2.9.0
numpy==1.16.3
pkg-resources==0.0.0
pycparser==2.19
python-telegram-bot==12.0.0b1
PyYAML==5.1
six==1.12.0
torch==1.1.0
tornado==6.0.2
